refactor(nexus-web): migrate Files UI to Account-style module
Split files browse into a colocated module with semantic CSS, mobile
toolbar/rows, and drive list vs browse routes for mobile list-detail.
