+++ libs/naas-abi/naas_abi/apps/nexus/apps/web/src/app/workspace/[workspaceId]/files/page.tsx
@@ -1,1984 +1,15 @@
 'use client';
 
-import { useEffect, useState, useRef, useCallback } from 'react';
-import { useRouter, useParams } from 'next/navigation';
-import ReactMarkdown from 'react-markdown';
-import remarkGfm from 'remark-gfm';
-import { Header } from '@/components/shell/header';
-import { 
-  FileCode, 
-  FileCode2,
-  FileText,
-  FileJson,
-  FileImage,
-  FileSpreadsheet,
-  FileArchive,
-  FileVideo,
-  FileAudio,
-  Presentation,
-  Folder, 
-  FolderPlus, 
-  RefreshCw, 
-  Upload, 
-  Download, 
-  Trash2,
-  MoreVertical,
-  Grid,
-  List,
-  Search,
-  FlaskConical,
-  Eye,
-  Code,
-  Star,
-  X,
-  ArrowUp,
-  ArrowDown,
-} from 'lucide-react';
-import { cn } from '@/lib/utils';
-import {
-  useFilesStore,
-  type FileInfo,
-  type FileSortKey,
-  type SortDirection,
-} from '@/stores/files';
-import { authFetch, useAuthStore } from '@/stores/auth';
-import { usePrompt, useConfirm } from '@/components/ui/dialogs';
-import { PdfViewer } from '@/components/files/pdf-viewer';
-
-type FileWithRelativeDir = { file: File; relativeDir?: string };
-
-async function readDirectoryEntries(
-  reader: FileSystemDirectoryReader,
-): Promise<FileSystemEntry[]> {
-  const all: FileSystemEntry[] = [];
-  while (true) {
-    const batch: FileSystemEntry[] = await new Promise((resolve, reject) => {
-      reader.readEntries(resolve, reject);
-    });
-    if (batch.length === 0) break;
-    all.push(...batch);
-  }
-  return all;
-}
-
-async function collectEntries(
-  entries: FileSystemEntry[],
-  parentDir = '',
-): Promise<FileWithRelativeDir[]> {
-  const out: FileWithRelativeDir[] = [];
-  for (const entry of entries) {
-    if (entry.isFile) {
-      const fileEntry = entry as FileSystemFileEntry;
-      const file = await new Promise<File>((resolve, reject) => {
-        fileEntry.file(resolve, reject);
-      });
-      out.push({ file, relativeDir: parentDir || undefined });
-    } else if (entry.isDirectory) {
-      const dirEntry = entry as FileSystemDirectoryEntry;
-      const children = await readDirectoryEntries(dirEntry.createReader());
-      const childDir = parentDir ? `${parentDir}/${entry.name}` : entry.name;
-      const nested = await collectEntries(children, childDir);
-      out.push(...nested);
-    }
-  }
-  return out;
-}
-
-// Client-side sort for local synced folders (loaded whole in the browser).
-// Mirrors the server ordering: case-insensitive name, numeric size, timestamp
-// modified — with missing size/modified sorting first on ascending order.
-function sortFiles(
-  files: FileInfo[],
-  sortBy: FileSortKey,
-  sortDir: SortDirection,
-): FileInfo[] {
-  const dir = sortDir === 'desc' ? -1 : 1;
-  return [...files].sort((a, b) => {
-    let cmp: number;
-    if (sortBy === 'size') {
-      cmp = (a.size ?? -1) - (b.size ?? -1);
-    } else if (sortBy === 'modified') {
-      const am = a.modified ? Date.parse(a.modified) : Number.NEGATIVE_INFINITY;
-      const bm = b.modified ? Date.parse(b.modified) : Number.NEGATIVE_INFINITY;
-      cmp = am - bm;
-    } else {
-      cmp = a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-    }
-    // Stable tiebreak by name so equal sizes/dates keep a predictable order.
-    if (cmp === 0 && sortBy !== 'name') {
-      cmp = a.name.toLowerCase().localeCompare(b.name.toLowerCase());
-    }
-    return cmp * dir;
-  });
-}
-
-export default function FilesPage() {
-  const router = useRouter();
-  const params = useParams();
-  const workspaceId = params.workspaceId as string;
-  
-  const {
-    files,
-    loading,
-    error,
-    currentPath,
-    filesTotal,
-    filesSortBy,
-    filesSortDir,
-    setFilesSort,
-    fetchFiles,
-    createFile,
-    createFolder,
-    deleteFile,
-    renameFile,
-    uploadFile,
-    uploadFiles,
-    refreshFiles,
-    openFile,
-    activeSource,
-    setActiveSource,
-    syncedFolders,
-    fetchLocalFiles,
-    setError,
-    uploadProgress,
-    starredItems,
-    starItem,
-    unstarItem,
-    starredNavigation,
-    setStarredNavigation,
-  } = useFilesStore();
-
-  const { prompt, dialog: promptDialog } = usePrompt();
-  const { confirm, dialog: confirmDialog } = useConfirm();
-
-  const [viewMode, setViewMode] = useState<'list' | 'grid'>('list');
-  const [searchQuery, setSearchQuery] = useState('');
-  // Debounced search term used to drive the server-side search fetch, so we
-  // don't hit the API on every keystroke.
-  const [debouncedSearch, setDebouncedSearch] = useState('');
-  useEffect(() => {
-    const timer = setTimeout(() => setDebouncedSearch(searchQuery.trim()), 300);
-    return () => clearTimeout(timer);
-  }, [searchQuery]);
-  // Client-side pagination: cap how many rows render at once so large folders
-  // (e.g. 200+ files) don't freeze the page while the API returns everything.
-  const PAGE_SIZE_OPTIONS = [10, 20, 50] as const;
-  const [pageSize, setPageSize] = useState<number>(50);
-  const [currentPage, setCurrentPage] = useState(1);
-  // Latest page size, readable from effects without making them re-run/re-fetch.
-  const pageSizeRef = useRef(pageSize);
-  useEffect(() => {
-    pageSizeRef.current = pageSize;
-  }, [pageSize]);
-  // The search term of the last fetch we issued. Updated synchronously so the
-  // debounced-search effect can tell a user edit apart from a fetch that
-  // navigation/mount already performed, avoiding a duplicate round-trip.
-  const lastSearchRef = useRef('');
-  const [selectedFiles, setSelectedFiles] = useState<string[]>([]);
-  const [isDragging, setIsDragging] = useState(false);
-  const [dropTargetPath, setDropTargetPath] = useState<string | null>(null);
-  const [activeContextMenu, setActiveContextMenu] = useState<string | null>(null);
-  const [pdfViewerFileName, setPdfViewerFileName] = useState<string | null>(null);
-  const [pdfViewerUrl, setPdfViewerUrl] = useState<string | null>(null);
-  const [pdfViewerLoading, setPdfViewerLoading] = useState(false);
-  const [pdfViewerError, setPdfViewerError] = useState<string | null>(null);
-  const [pdfViewerDownloadFile, setPdfViewerDownloadFile] = useState<FileInfo | null>(null);
-  const [imageViewerFileName, setImageViewerFileName] = useState<string | null>(null);
-  const [imageViewerUrl, setImageViewerUrl] = useState<string | null>(null);
-  const [imageViewerLoading, setImageViewerLoading] = useState(false);
-  const [imageViewerError, setImageViewerError] = useState<string | null>(null);
-  const [textViewerFileName, setTextViewerFileName] = useState<string | null>(null);
-  const [textViewerContent, setTextViewerContent] = useState<string>('');
-  const [textViewerMode, setTextViewerMode] = useState<'markdown' | 'code'>('code');
-  // HTML files get a rendered preview with a toggle between preview and source.
-  const [textViewerIsHtml, setTextViewerIsHtml] = useState(false);
-  const [htmlPreview, setHtmlPreview] = useState(true);
-  const [textViewerLoading, setTextViewerLoading] = useState(false);
-  const [textViewerError, setTextViewerError] = useState<string | null>(null);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  // When a starred navigation is pending, skip the regular fetch triggered by activeSource change
-  const skipRegularFetchRef = useRef(false);
-  // Local path of the file to auto-preview once files finish loading
-  const [localPendingPreview, setLocalPendingPreview] = useState<string | null>(null);
-
-  // Helper to open file in Lab
-  const handleOpenInLab = (file: FileInfo) => {
-    openFile(file.path);
-    router.push(`/workspace/${workspaceId}/lab`);
-  };
-  
-  // Close context menu when clicking outside
-  useEffect(() => {
-    const handleClickOutside = () => setActiveContextMenu(null);
-    if (activeContextMenu) {
-      document.addEventListener('click', handleClickOutside);
-      return () => document.removeEventListener('click', handleClickOutside);
-    }
-  }, [activeContextMenu]);
-
-  useEffect(() => {
-    return () => {
-      if (pdfViewerUrl) {
-        URL.revokeObjectURL(pdfViewerUrl);
-      }
-    };
-  }, [pdfViewerUrl]);
-
-  // Close any open preview viewer when the user presses Escape.
-  useEffect(() => {
-    if (!pdfViewerFileName && !imageViewerFileName && !textViewerFileName) return;
-    const handleKeyDown = (e: KeyboardEvent) => {
-      if (e.key !== 'Escape') return;
-      if (pdfViewerFileName) closePdfViewer();
-      else if (imageViewerFileName) closeImageViewer();
-      else if (textViewerFileName) closeTextViewer();
-    };
-    document.addEventListener('keydown', handleKeyDown);
-    return () => document.removeEventListener('keydown', handleKeyDown);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pdfViewerFileName, imageViewerFileName, textViewerFileName]);
-
-  useEffect(() => {
-    return () => {
-      if (imageViewerUrl) {
-        URL.revokeObjectURL(imageViewerUrl);
-      }
-    };
-  }, [imageViewerUrl]);
-
-  // Handle navigation triggered by clicking a starred item in the sidebar.
-  // Setting skipRegularFetchRef prevents the activeSource-change effect from
-  // overriding the path we set here.
-  useEffect(() => {
-    if (!starredNavigation) return;
-    skipRegularFetchRef.current = true;
-    setActiveSource(starredNavigation.source);
-    setSearchQuery('');
-    setDebouncedSearch('');
-    lastSearchRef.current = '';
-    setCurrentPage(1);
-    fetchFiles(starredNavigation.path, { limit: pageSizeRef.current, offset: 0, search: '' });
-    if (starredNavigation.previewPath) {
-      setLocalPendingPreview(starredNavigation.previewPath);
-    }
-    setStarredNavigation(null);
-  }, [starredNavigation, setActiveSource, fetchFiles, setStarredNavigation]);
-
-  // Once the files list loads, open the pending preview (if any).
-  useEffect(() => {
-    if (!localPendingPreview || loading) return;
-    const target = files.find((f) => f.path === localPendingPreview);
-    if (!target) return;
-    setLocalPendingPreview(null);
-    if (isPdfFile(target)) openPdfViewer(target);
-    else if (isOfficeFile(target)) openOfficePreview(target);
-    else if (isImageFile(target)) openImageViewer(target);
-    else if (isTextFile(target)) openTextViewer(target);
-  // isPdfFile etc. are stable inline functions — listing them would cause an infinite loop
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [localPendingPreview, files, loading]);
-  
-  // Check if we're viewing a synced local folder
-  const activeSyncedFolder = syncedFolders.find((f) => f.id === activeSource);
-  const isLocalFolder = !!activeSyncedFolder;
-  // Source-aware scope param to match how file ops are routed in the store.
-  const filesScope: 'workspace' | 'my_drive' | 'platform_drive' | 'system_drive' =
-    activeSource === 'my-drive'
-      ? 'my_drive'
-      : activeSource === 'platform-drive'
-        ? 'platform_drive'
-        : activeSource === 'system-drive'
-          ? 'system_drive'
-          : 'workspace';
-  const fileQueryParams = `workspace_id=${encodeURIComponent(workspaceId)}&scope=${filesScope}`;
-
-  // Compute the drive's storage root and a user-facing label. The breadcrumb
-  // hides anything above the drive — users have no business navigating into
-  // naas_abi/<drive>/<entity-id>/... above their drive.
-  const authUserId = useAuthStore((state) => state.user?.id);
-  const driveRoot =
-    activeSource === 'my-drive'
-      ? authUserId
-        ? `naas_abi/my-drive/${authUserId}`
-        : ''
-      : activeSource === 'platform-drive'
-        ? 'naas_abi/platform-drive'
-        : activeSource === 'system-drive'
-          ? ''
-          : workspaceId
-            ? `naas_abi/workspace-drive/${workspaceId}`
-            : '';
-  const driveLabel = isLocalFolder
-    ? activeSyncedFolder?.name || 'Drive'
-    : activeSource === 'my-drive'
-      ? 'My Drive'
-      : activeSource === 'platform-drive'
-        ? 'Platform Drive'
-        : activeSource === 'system-drive'
-          ? 'System Drive'
-          : 'Workspace Drive';
-
-  // Strip the drive root (and the legacy bare-workspace-id root, for backwards
-  // compatibility) from the current path so the breadcrumb only shows what's
-  // under the drive.
-  const stripRoot = (path: string, root: string): string | null => {
-    const normalized = path.replace(/^\/+|\/+$/g, '');
-    if (!root) return null;
-    if (normalized === root) return '';
-    if (normalized.startsWith(`${root}/`)) return normalized.slice(root.length + 1);
-    return null;
-  };
-  const legacyRoots = activeSource === 'workspace' && workspaceId ? [workspaceId] : [];
-  let relativePath = (currentPath || '').replace(/^\/+|\/+$/g, '');
-  for (const root of [driveRoot, ...legacyRoots]) {
-    const stripped = stripRoot(relativePath, root);
-    if (stripped !== null) {
-      relativePath = stripped;
-      break;
-    }
-  }
-
-  useEffect(() => {
-    // Skip if a starred navigation just updated activeSource — it already called fetchFiles.
-    if (skipRegularFetchRef.current) {
-      skipRegularFetchRef.current = false;
-      return;
-    }
-
-    // Clear any previous errors
-    setError(null);
-
-    // Fetch files based on active source. Remote drives fetch the first page;
-    // local synced folders load fully and paginate client-side.
-    if (isLocalFolder && activeSyncedFolder) {
-      fetchLocalFiles(activeSyncedFolder.id);
-    } else {
-      setSearchQuery('');
-      setDebouncedSearch('');
-      lastSearchRef.current = '';
-      setCurrentPage(1);
-      fetchFiles('', { limit: pageSizeRef.current, offset: 0, search: '' });
-    }
-  }, [fetchFiles, fetchLocalFiles, isLocalFolder, activeSyncedFolder, activeSource, setError]);
-
-  // Drop selections that no longer exist (after navigation or deletion).
-  useEffect(() => {
-    setSelectedFiles((prev) => prev.filter((p) => files.some((f) => f.path === p)));
-  }, [files]);
-
-  // Warn the user before leaving while an upload is in progress.
-  useEffect(() => {
-    if (!uploadProgress?.active) return;
-    const handler = (e: BeforeUnloadEvent) => {
-      e.preventDefault();
-      e.returnValue = '';
-    };
-    window.addEventListener('beforeunload', handler);
-    return () => window.removeEventListener('beforeunload', handler);
-  }, [uploadProgress?.active]);
-
-  const INTERNAL_DRAG_MIME = 'application/x-nexus-file';
-
-  const isInternalDrag = (e: React.DragEvent) =>
-    e.dataTransfer.types.includes(INTERNAL_DRAG_MIME);
-
-  // Drag and drop handlers (external upload only — internal moves bubble up but are ignored here)
-  const handleDragOver = useCallback((e: React.DragEvent) => {
-    if (isInternalDrag(e)) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(true);
-  }, []);
-
-  const handleDragLeave = useCallback((e: React.DragEvent) => {
-    if (isInternalDrag(e)) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-  }, []);
-
-  const handleDrop = useCallback(async (e: React.DragEvent) => {
-    if (isInternalDrag(e)) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setIsDragging(false);
-
-    const items = e.dataTransfer.items;
-    if (items && items.length > 0 && typeof items[0].webkitGetAsEntry === 'function') {
-      const entries: FileSystemEntry[] = [];
-      for (let i = 0; i < items.length; i++) {
-        const entry = items[i].webkitGetAsEntry();
-        if (entry) entries.push(entry);
-      }
-      if (entries.length > 0) {
-        const collected = await collectEntries(entries);
-        if (collected.length > 0) {
-          await uploadFiles(collected);
-        }
-        return;
-      }
-    }
-
-    const droppedFiles = e.dataTransfer.files;
-    if (droppedFiles.length > 0) {
-      await uploadFiles(droppedFiles);
-    }
-  }, [uploadFiles]);
-
-  // Internal move via drag onto a folder row.
-  const handleRowDragStart = (e: React.DragEvent, file: FileInfo) => {
-    e.dataTransfer.setData(INTERNAL_DRAG_MIME, file.path);
-    e.dataTransfer.effectAllowed = 'move';
-  };
-
-  const handleFolderDragOver = (e: React.DragEvent, folder: FileInfo) => {
-    if (!isInternalDrag(e)) return;
-    e.preventDefault();
-    e.stopPropagation();
-    e.dataTransfer.dropEffect = 'move';
-    setDropTargetPath(folder.path);
-  };
-
-  const handleFolderDragLeave = (e: React.DragEvent, folder: FileInfo) => {
-    if (!isInternalDrag(e)) return;
-    e.stopPropagation();
-    setDropTargetPath((current) => (current === folder.path ? null : current));
-  };
-
-  const handleFolderDrop = async (e: React.DragEvent, folder: FileInfo) => {
-    if (!isInternalDrag(e)) return;
-    e.preventDefault();
-    e.stopPropagation();
-    setDropTargetPath(null);
-
-    const sourcePath = e.dataTransfer.getData(INTERNAL_DRAG_MIME);
-    if (!sourcePath) return;
-    // dataTransfer.getData isn't always available during dragover, so re-validate here
-    const sourceParent = sourcePath.includes('/') ? sourcePath.slice(0, sourcePath.lastIndexOf('/')) : '';
-    if (
-      sourcePath === folder.path ||
-      sourceParent === folder.path ||
-      folder.path === sourcePath ||
-      folder.path.startsWith(`${sourcePath}/`)
-    ) {
-      return;
-    }
-    const name = sourcePath.includes('/') ? sourcePath.slice(sourcePath.lastIndexOf('/') + 1) : sourcePath;
-    const newPath = `${folder.path}/${name}`;
-    await renameFile(sourcePath, newPath);
-  };
-
-  const handleFileInputChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedFiles = e.target.files;
-    if (selectedFiles && selectedFiles.length > 0) {
-      await uploadFiles(selectedFiles);
-      // Reset input
-      if (fileInputRef.current) {
-        fileInputRef.current.value = '';
-      }
-    }
-  };
-
-  // Generate unique name if the proposed name already exists
-  const getUniqueName = (baseName: string, extension: string = '') => {
-    const existingNames = new Set(files.map((f) => f.name.toLowerCase()));
-    let name = extension ? `${baseName}${extension}` : baseName;
-    let counter = 1;
-    
-    while (existingNames.has(name.toLowerCase())) {
-      name = extension ? `${baseName}-${counter}${extension}` : `${baseName}-${counter}`;
-      counter++;
-    }
-    return name;
-  };
-
-  const handleNewFile = async () => {
-    const defaultName = getUniqueName('untitled', '.txt');
-    const name = await prompt({
-      title: 'New File',
-      description: 'Enter a name for the new file',
-      defaultValue: defaultName,
-      confirmLabel: 'Create',
-    });
-    if (name) {
-      const fullPath = currentPath ? `${currentPath}/${name}` : name;
-      await createFile(fullPath);
-    }
-  };
-
-  const handleNewFolder = async () => {
-    const defaultName = getUniqueName('new-folder');
-    const name = await prompt({
-      title: 'New Folder',
-      description: 'Enter a name for the new folder',
-      defaultValue: defaultName,
-      confirmLabel: 'Create',
-    });
-    if (name) {
-      const fullPath = currentPath ? `${currentPath}/${name}` : name;
-      await createFolder(fullPath);
-    }
-  };
-
-  const handleRename = async (file: FileInfo) => {
-    const existingNames = new Set(files.map((f) => f.name.toLowerCase()));
-    const newName = await prompt({
-      title: 'Rename',
-      description: `Rename "${file.name}" to:`,
-      defaultValue: file.name,
-      confirmLabel: 'Rename',
-    });
-    if (newName && newName !== file.name) {
-      if (existingNames.has(newName.toLowerCase()) && newName.toLowerCase() !== file.name.toLowerCase()) {
-        // Show a second prompt with the error -- re-trigger rename
-        await prompt({
-          title: 'Name already exists',
-          description: `A file or folder named "${newName}" already exists. Choose a different name.`,
-          defaultValue: newName,
-          confirmLabel: 'Rename',
-        }).then(async (retryName) => {
-          if (retryName && retryName !== file.name) {
-            const parentPath = file.path.substring(0, file.path.lastIndexOf('/'));
-            const newPath = parentPath ? `${parentPath}/${retryName}` : retryName;
-            await renameFile(file.path, newPath);
-          }
-        });
-        return;
-      }
-
-      const parentPath = file.path.substring(0, file.path.lastIndexOf('/'));
-      const newPath = parentPath ? `${parentPath}/${newName}` : newName;
-      await renameFile(file.path, newPath);
-    }
-  };
-
-  const handleDelete = async (file: FileInfo) => {
-    const confirmed = await confirm({
-      title: `Delete "${file.name}"?`,
-      description: file.type === 'folder'
-        ? 'This folder and all its contents will be permanently deleted.'
-        : 'This file will be permanently deleted.',
-      confirmLabel: 'Delete',
-      destructive: true,
-    });
-    if (confirmed) {
-      await deleteFile(file.path);
-    }
-  };
-
-  const handleDeleteSelected = async () => {
-    if (selectedFiles.length === 0) return;
-    const count = selectedFiles.length;
-    const confirmed = await confirm({
-      title: `Delete ${count} item${count === 1 ? '' : 's'}?`,
-      description: 'The selected files and folders will be permanently deleted.',
-      confirmLabel: 'Delete',
-      destructive: true,
-    });
-    if (!confirmed) return;
-    // Delete sequentially so a single failure doesn't abandon the rest.
-    for (const path of selectedFiles) {
-      await deleteFile(path);
-    }
-    setSelectedFiles([]);
-  };
-
-  // Remote drives paginate on the server (the store fetches one page at a time),
-  // so `files` already holds just the current page. Local synced folders are read
-  // entirely in the browser via the File System Access API, so they paginate
-  // client-side by slicing the loaded list.
-  const isServerPaginated = !isLocalFolder;
-
-  // Remote search + sort are applied by the server, so `files` is already the
-  // filtered, sorted page — use it as-is. Local folders are loaded whole in the
-  // browser, so we filter and sort them client-side to match.
-  const filteredFiles = isServerPaginated
-    ? files
-    : sortFiles(
-        files.filter((file) =>
-          file.name.toLowerCase().includes(searchQuery.toLowerCase())
-        ),
-        filesSortBy,
-        filesSortDir,
-      );
-  const pageCount = isServerPaginated ? filesTotal : filteredFiles.length;
-  const totalPages = Math.max(1, Math.ceil(pageCount / pageSize));
-  const safePage = Math.min(currentPage, totalPages);
-  const pageStart = (safePage - 1) * pageSize;
-  const pagedFiles = isServerPaginated
-    ? filteredFiles
-    : filteredFiles.slice(pageStart, pageStart + pageSize);
-  // Range shown in the pagination footer ("X–Y of N").
-  const rangeStart = pageCount === 0 ? 0 : pageStart + 1;
-  const rangeEnd = isServerPaginated
-    ? pageStart + pagedFiles.length
-    : Math.min(pageStart + pageSize, filteredFiles.length);
-
-  // Fetch a page from the server (remote drives only). Local folders just move
-  // the client-side window via setCurrentPage.
-  const goToPage = useCallback(
-    (page: number) => {
-      const clamped = Math.max(1, page);
-      setCurrentPage(clamped);
-      if (isServerPaginated) {
-        fetchFiles(currentPath, {
-          limit: pageSize,
-          offset: (clamped - 1) * pageSize,
-          search: debouncedSearch,
-        });
-      }
-    },
-    [isServerPaginated, fetchFiles, currentPath, pageSize, debouncedSearch],
-  );
-
-  const changePageSize = (size: number) => {
-    setPageSize(size);
-    setCurrentPage(1);
-    if (isServerPaginated) {
-      fetchFiles(currentPath, { limit: size, offset: 0, search: debouncedSearch });
-    }
-  };
-
-  // Sort a column. setFilesSort updates the persisted preference synchronously,
-  // so the follow-up fetch (which reads sort from the store) uses the new order.
-  // Local folders re-sort on render, no fetch needed.
-  const handleSort = (column: FileSortKey) => {
-    setFilesSort(column);
-    setCurrentPage(1);
-    if (isServerPaginated) {
-      fetchFiles(currentPath, { limit: pageSize, offset: 0, search: debouncedSearch });
-    }
-  };
-
-  // Header cell that sorts its column, with an active-direction arrow.
-  const SortHeader = ({ column, label }: { column: FileSortKey; label: string }) => {
-    const active = filesSortBy === column;
-    return (
-      <button
-        onClick={() => handleSort(column)}
-        className={cn(
-          'flex items-center gap-1 font-medium hover:text-foreground',
-          active ? 'text-foreground' : 'text-muted-foreground',
-        )}
-      >
-        {label}
-        {active &&
-          (filesSortDir === 'asc' ? <ArrowUp size={12} /> : <ArrowDown size={12} />)}
-      </button>
-    );
-  };
-
-  // Navigate into a folder / breadcrumb target: clear any search, reset to page 1
-  // and fetch the first server page. Used by the JSX click handlers below.
-  const navigateToPath = useCallback(
-    (path: string) => {
-      setSearchQuery('');
-      setDebouncedSearch('');
-      lastSearchRef.current = '';
-      setCurrentPage(1);
-      fetchFiles(path, { limit: pageSizeRef.current, offset: 0, search: '' });
-    },
-    [fetchFiles],
-  );
-
-  // Reset to the first page whenever the search box changes (both modes).
-  useEffect(() => {
-    setCurrentPage(1);
-  }, [searchQuery]);
-
-  // Server-side search: refetch the first page when the debounced term changes.
-  // Guarded against the last-fetched term so navigation/mount fetches (which
-  // already carry the right search) don't trigger a redundant round-trip.
-  useEffect(() => {
-    if (!isServerPaginated) return;
-    if (debouncedSearch === lastSearchRef.current) return;
-    lastSearchRef.current = debouncedSearch;
-    setCurrentPage(1);
-    fetchFiles(currentPath, { limit: pageSizeRef.current, offset: 0, search: debouncedSearch });
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [debouncedSearch]);
-
-  // After the total shrinks (e.g. deleting the last item on the last page), pull
-  // the current page back into range and reload it. Server-paginated only.
-  useEffect(() => {
-    if (!isServerPaginated) return;
-    const maxPage = Math.max(1, Math.ceil(filesTotal / pageSize));
-    if (currentPage > maxPage) {
-      setCurrentPage(maxPage);
-      fetchFiles(currentPath, {
-        limit: pageSize,
-        offset: (maxPage - 1) * pageSize,
-        search: lastSearchRef.current,
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [filesTotal]);
-
-  const toggleSelectFile = (path: string) => {
-    setSelectedFiles((prev) =>
-      prev.includes(path) ? prev.filter((p) => p !== path) : [...prev, path]
-    );
-  };
-
-  // Select-all operates on the currently visible page so the header checkbox
-  // reflects what the user can actually see.
-  const allSelected =
-    pagedFiles.length > 0 && pagedFiles.every((f) => selectedFiles.includes(f.path));
-  const someSelected = selectedFiles.length > 0 && !allSelected;
-
-  const toggleSelectAll = () => {
-    const pagePaths = pagedFiles.map((f) => f.path);
-    setSelectedFiles((prev) =>
-      allSelected
-        ? prev.filter((p) => !pagePaths.includes(p))
-        : Array.from(new Set([...prev, ...pagePaths]))
-    );
-  };
-
-  const formatSize = (bytes?: number) => {
-    if (!bytes) return '—';
-    if (bytes < 1024) return `${bytes} B`;
-    if (bytes < 1024 * 1024) return `${(bytes / 1024).toFixed(1)} KB`;
-    return `${(bytes / (1024 * 1024)).toFixed(1)} MB`;
-  };
-
-  const formatDate = (date?: string) => {
-    if (!date) return '—';
-    return new Date(date).toLocaleDateString('en-US', {
-      month: 'short',
-      day: 'numeric',
-      year: 'numeric',
-    });
-  };
-
-  const getFileExtension = (name: string) => {
-    const idx = name.lastIndexOf('.');
-    return idx >= 0 ? name.slice(idx + 1).toLowerCase() : '';
-  };
-
-  const getFileIcon = (file: FileInfo, size = 16) => {
-    if (file.type === 'folder') {
-      return <Folder size={size} className="text-muted-foreground" />;
-    }
-
-    const ext = getFileExtension(file.name);
-    const contentType = file.content_type?.toLowerCase() || '';
-
-    if (ext === 'pdf') return <FileText size={size} className="text-red-500" />;
-    if (['md', 'markdown'].includes(ext)) return <FileText size={size} className="text-sky-600" />;
-    if (['ppt', 'pptx', 'key'].includes(ext)) return <Presentation size={size} className="text-orange-500" />;
-    if (['txt', 'doc', 'docx', 'rtf'].includes(ext)) return <FileText size={size} className="text-blue-500" />;
-    if (['xls', 'xlsx', 'xlsm', 'xlsb'].includes(ext)) return <FileSpreadsheet size={size} className="text-emerald-700" />;
-    if (['csv', 'tsv'].includes(ext)) return <FileSpreadsheet size={size} className="text-emerald-500" />;
-    if (['json', 'yaml', 'yml'].includes(ext)) return <FileJson size={size} className="text-amber-500" />;
-    if (['zip', 'rar', '7z', 'tar', 'gz'].includes(ext)) return <FileArchive size={size} className="text-orange-500" />;
-    if (['mp4', 'mov', 'avi', 'mkv', 'webm'].includes(ext) || contentType.startsWith('video/')) {
-      return <FileVideo size={size} className="text-fuchsia-500" />;
-    }
-    if (['mp3', 'wav', 'ogg', 'flac', 'm4a'].includes(ext) || contentType.startsWith('audio/')) {
-      return <FileAudio size={size} className="text-pink-500" />;
-    }
-    if (
-      ['png', 'jpg', 'jpeg', 'gif', 'webp', 'svg'].includes(ext) ||
-      contentType.startsWith('image/')
-    ) {
-      return <FileImage size={size} className="text-violet-500" />;
-    }
-    if (
-      ['js', 'jsx', 'ts', 'tsx', 'py', 'java', 'go', 'rs', 'rb', 'php', 'cs', 'cpp', 'c', 'h', 'sh', 'sql'].includes(ext)
-    ) {
-      return <FileCode2 size={size} className="text-cyan-500" />;
-    }
-
-    return <FileCode size={size} className="text-muted-foreground" />;
-  };
-
-  const isPdfFile = (file: FileInfo) =>
-    file.type === 'file' && (
-      file.name.toLowerCase().endsWith('.pdf') ||
-      file.content_type === 'application/pdf'
-    );
-
-  const isImageFile = (file: FileInfo) => {
-    if (file.type !== 'file') return false;
-    const lowerName = file.name.toLowerCase();
-    return (
-      lowerName.endsWith('.png') ||
-      lowerName.endsWith('.jpg') ||
-      lowerName.endsWith('.jpeg') ||
-      lowerName.endsWith('.gif') ||
-      lowerName.endsWith('.webp') ||
-      lowerName.endsWith('.svg') ||
-      file.content_type?.startsWith('image/') === true
-    );
-  };
-
-  const isHtmlFile = (file: FileInfo) => {
-    if (file.type !== 'file') return false;
-    const lowerName = file.name.toLowerCase();
-    return (
-      lowerName.endsWith('.html') ||
-      lowerName.endsWith('.htm') ||
-      file.content_type === 'text/html'
-    );
-  };
-
-  const isMarkdownFile = (file: FileInfo) => {
-    if (file.type !== 'file') return false;
-    const lowerName = file.name.toLowerCase();
-    return (
-      lowerName.endsWith('.md') ||
-      lowerName.endsWith('.markdown') ||
-      lowerName.endsWith('.mdx') ||
-      file.content_type === 'text/markdown'
-    );
-  };
-
-  // Extensions that are reliably plain text. Markdown is detected separately
-  // so it can be rendered with ReactMarkdown instead of as raw source.
-  const TEXT_EXTENSIONS = new Set([
-    'txt', 'log', 'rst', 'tex', 'csv', 'tsv',
-    'json', 'jsonc', 'json5', 'ndjson', 'geojson',
-    'yaml', 'yml', 'toml', 'ini', 'cfg', 'conf', 'properties', 'env',
-    'xml', 'plist', 'svg', 'gql', 'graphql', 'proto',
-    'html', 'htm', 'css', 'scss', 'sass', 'less', 'styl',
-    'js', 'jsx', 'mjs', 'cjs', 'ts', 'tsx', 'vue', 'svelte', 'astro',
-    'py', 'pyi', 'rb', 'go', 'rs', 'java', 'kt', 'kts', 'swift',
-    'c', 'h', 'cc', 'cpp', 'hpp', 'cs', 'm', 'mm',
-    'php', 'lua', 'r', 'jl', 'dart', 'sql', 'pl', 'pm',
-    'sh', 'bash', 'zsh', 'fish', 'ps1', 'bat', 'cmd',
-    'patch', 'diff', 'gitignore', 'gitattributes', 'editorconfig',
-    'lock', 'in', 'mk',
-  ]);
-
-  // Filenames (no extension or special) that we know are text.
-  const TEXT_FILENAMES = new Set([
-    'dockerfile', 'makefile', 'rakefile', 'gemfile', 'procfile',
-    'license', 'readme', 'changelog', 'authors', 'contributors', 'notice',
-    '.gitignore', '.gitattributes', '.editorconfig', '.env', '.npmrc', '.nvmrc',
-  ]);
-
-  const isTextFile = (file: FileInfo) => {
-    if (file.type !== 'file') return false;
-    if (isMarkdownFile(file)) return true;
-    const lowerName = file.name.toLowerCase();
-    if (TEXT_FILENAMES.has(lowerName)) return true;
-    const ext = getFileExtension(file.name);
-    if (ext && TEXT_EXTENSIONS.has(ext)) return true;
-    const ct = file.content_type?.toLowerCase() || '';
-    if (ct.startsWith('text/')) return true;
-    if (
-      ct === 'application/json' ||
-      ct === 'application/xml' ||
-      ct === 'application/javascript' ||
-      ct === 'application/x-yaml' ||
-      ct === 'application/x-sh' ||
-      ct.endsWith('+json') ||
-      ct.endsWith('+xml')
-    ) {
-      return true;
-    }
-    return false;
-  };
-
-  // Office formats we render via server-side LibreOffice → PDF conversion.
-  const OFFICE_PREVIEW_EXTENSIONS = new Set([
-    'ppt', 'pptx', 'odp', 'key',
-    'doc', 'docx', 'odt', 'rtf',
-    'xls', 'xlsx', 'xlsm', 'xlsb', 'ods',
-  ]);
-
-  const isOfficeFile = (file: FileInfo) => {
-    if (file.type !== 'file') return false;
-    return OFFICE_PREVIEW_EXTENSIONS.has(getFileExtension(file.name));
-  };
-
-  const officeKindLabel = (file: FileInfo): string => {
-    const ext = getFileExtension(file.name);
-    if (['doc', 'docx', 'odt', 'rtf'].includes(ext)) return 'Document';
-    if (['xls', 'xlsx', 'xlsm', 'xlsb', 'ods'].includes(ext)) return 'Spreadsheet';
-    return 'Presentation';
-  };
-
-  const openPdfViewer = async (file: FileInfo) => {
-    if (!isPdfFile(file)) return;
-    setPdfViewerError(null);
-    setPdfViewerLoading(true);
-    setPdfViewerFileName(file.name);
-    setPdfViewerDownloadFile(null);
-
-    if (pdfViewerUrl) {
-      URL.revokeObjectURL(pdfViewerUrl);
-      setPdfViewerUrl(null);
-    }
-
-    try {
-      const encodedPath = file.path.split('/').map(encodeURIComponent).join('/');
-      const response = await authFetch(
-        `/api/files/raw/${encodedPath}?${fileQueryParams}`
-      );
-      if (!response.ok) {
-        throw new Error('Failed to load PDF preview');
-      }
-      const blob = await response.blob();
-      const blobUrl = URL.createObjectURL(
-        blob.type ? blob : new Blob([blob], { type: 'application/pdf' })
-      );
-      setPdfViewerUrl(blobUrl);
-    } catch (error) {
-      setPdfViewerError(error instanceof Error ? error.message : 'Failed to load PDF preview');
-    } finally {
-      setPdfViewerLoading(false);
-    }
-  };
-
-  const closePdfViewer = () => {
-    setPdfViewerFileName(null);
-    setPdfViewerError(null);
-    setPdfViewerDownloadFile(null);
-    if (pdfViewerUrl) {
-      URL.revokeObjectURL(pdfViewerUrl);
-    }
-    setPdfViewerUrl(null);
-    setPdfViewerLoading(false);
-  };
-
-  const openOfficePreview = async (file: FileInfo) => {
-    if (!isOfficeFile(file)) return;
-    setPdfViewerError(null);
-    setPdfViewerLoading(true);
-    setPdfViewerFileName(`${file.name} (PDF preview)`);
-    setPdfViewerDownloadFile(file);
-
-    if (pdfViewerUrl) {
-      URL.revokeObjectURL(pdfViewerUrl);
-      setPdfViewerUrl(null);
-    }
-
-    try {
-      const encodedPath = file.path.split('/').map(encodeURIComponent).join('/');
-      const response = await authFetch(
-        `/api/files/preview/pdf/${encodedPath}?${fileQueryParams}`
-      );
-      if (!response.ok) {
-        throw new Error(`${officeKindLabel(file)} preview unavailable right now`);
-      }
-      const blob = await response.blob();
-      const blobUrl = URL.createObjectURL(
-        blob.type ? blob : new Blob([blob], { type: 'application/pdf' })
-      );
-      setPdfViewerUrl(blobUrl);
-    } catch (error) {
-      setPdfViewerFileName(null);
-      setPdfViewerDownloadFile(null);
-      setPdfViewerLoading(false);
-      // Graceful fallback: download original presentation when conversion is unavailable.
-      await downloadFileToDesktop(file);
-      return;
-    } finally {
-      setPdfViewerLoading(false);
-    }
-  };
-
-  const openImageViewer = async (file: FileInfo) => {
-    if (!isImageFile(file)) return;
-    setImageViewerError(null);
-    setImageViewerLoading(true);
-    setImageViewerFileName(file.name);
-
-    if (imageViewerUrl) {
-      URL.revokeObjectURL(imageViewerUrl);
-      setImageViewerUrl(null);
-    }
-
-    try {
-      const encodedPath = file.path.split('/').map(encodeURIComponent).join('/');
-      const response = await authFetch(
-        `/api/files/raw/${encodedPath}?${fileQueryParams}`
-      );
-      if (!response.ok) {
-        throw new Error('Failed to load image preview');
-      }
-      const blob = await response.blob();
-      const blobUrl = URL.createObjectURL(blob);
-      setImageViewerUrl(blobUrl);
-    } catch (error) {
-      setImageViewerError(error instanceof Error ? error.message : 'Failed to load image preview');
-    } finally {
-      setImageViewerLoading(false);
-    }
-  };
-
-  const closeImageViewer = () => {
-    setImageViewerFileName(null);
-    setImageViewerError(null);
-    if (imageViewerUrl) {
-      URL.revokeObjectURL(imageViewerUrl);
-    }
-    setImageViewerUrl(null);
-    setImageViewerLoading(false);
-  };
-
-  const openTextViewer = async (file: FileInfo) => {
-    if (!isTextFile(file)) return;
-    const mode: 'markdown' | 'code' = isMarkdownFile(file) ? 'markdown' : 'code';
-    const html = isHtmlFile(file);
-    setTextViewerError(null);
-    setTextViewerLoading(true);
-    setTextViewerFileName(file.name);
-    setTextViewerMode(mode);
-    setTextViewerIsHtml(html);
-    setHtmlPreview(html); // HTML opens rendered; user can toggle to source
-    setTextViewerContent('');
-
-    try {
-      const encodedPath = file.path.split('/').map(encodeURIComponent).join('/');
-      const response = await authFetch(
-        `/api/files/${encodedPath}?${fileQueryParams}`
-      );
-      if (!response.ok) {
-        if (response.status === 415 || response.status === 422) {
-          throw new Error('This file is not a UTF-8 text file.');
-        }
-        throw new Error('Failed to load preview');
-      }
-      const data = await response.json();
-      setTextViewerContent(typeof data.content === 'string' ? data.content : '');
-    } catch (error) {
-      setTextViewerError(error instanceof Error ? error.message : 'Failed to load preview');
-    } finally {
-      setTextViewerLoading(false);
-    }
-  };
-
-  const closeTextViewer = () => {
-    setTextViewerFileName(null);
-    setTextViewerContent('');
-    setTextViewerError(null);
-    setTextViewerLoading(false);
-    setTextViewerIsHtml(false);
-    setHtmlPreview(true);
-  };
-
-  const downloadFileToDesktop = async (file: FileInfo) => {
-    if (file.type !== 'file') return;
-    try {
-      const encodedPath = file.path.split('/').map(encodeURIComponent).join('/');
-      const response = await authFetch(
-        `/api/files/raw/${encodedPath}?${fileQueryParams}`
-      );
-      if (!response.ok) {
-        throw new Error('Failed to download file');
-      }
-      const blob = await response.blob();
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = file.name;
-      document.body.appendChild(a);
-      a.click();
-      a.remove();
-      URL.revokeObjectURL(url);
-    } catch (error) {
-      setError(error instanceof Error ? error.message : 'Failed to download file');
-    }
-  };
-
-  const downloadFolderToDesktop = async (folder: FileInfo) => {
-    if (folder.type !== 'folder') return;
-    try {
-      const encodedPath = folder.path.split('/').map(encodeURIComponent).join('/');
-      const response = await authFetch(
-        `/api/files/archive/${encodedPath}?${fileQueryParams}`
-      );
-      if (!response.ok) {
-        let detail = 'Failed to download folder';
-        try {
-          const data = await response.json();
-          if (data?.detail) detail = data.detail;
-        } catch {
-          /* ignore */
-        }
-        throw new Error(detail);
-      }
-      const blob = await response.blob();
-      const url = URL.createObjectURL(blob);
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = `${folder.name}.zip`;
-      document.body.appendChild(a);
-      a.click();
-      a.remove();
-      URL.revokeObjectURL(url);
-    } catch (error) {
-      setError(error instanceof Error ? error.message : 'Failed to download folder');
-    }
-  };
-
-  const handleRefresh = () => {
-    if (isLocalFolder && activeSyncedFolder) {
-      fetchLocalFiles(activeSyncedFolder.id, currentPath);
-    } else {
-      refreshFiles();
-    }
-  };
-
-  return (
-    <>
-    {promptDialog}
-    {confirmDialog}
-    <div className="flex h-full flex-col">
-      <Header 
-        title={isLocalFolder ? activeSyncedFolder?.name || 'Local Folder' : 'Files'} 
-        subtitle={isLocalFolder ? `Synced from your machine` : 'Manage workspace files and folders'} 
-      />
-
-      {/* Error banner */}
-      {error && (
-        <div className="flex items-center justify-between bg-destructive/10 px-4 py-2 text-sm text-destructive">
-          <span>{error}</span>
-          <button
-            onClick={() => setError(null)}
-            className="rounded px-2 py-0.5 hover:bg-destructive/20"
-          >
-            Dismiss
-          </button>
-        </div>
-      )}
-
-      <div className="flex flex-1 flex-col overflow-hidden">
-        {/* Breadcrumb (file path) — shown above the toolbar row */}
-        <div className="flex items-center gap-1 border-b px-4 py-2 text-sm">
-          <button
-            onClick={() => {
-              if (isLocalFolder && activeSyncedFolder) {
-                fetchLocalFiles(activeSyncedFolder.id, '');
-              } else {
-                navigateToPath('');
-              }
-            }}
-            className={cn(
-              relativePath ? 'text-muted-foreground hover:text-foreground' : 'text-foreground'
-            )}
-          >
-            {driveLabel}
-          </button>
-          {relativePath && relativePath.split('/').map((part, i, arr) => (
-            <span key={i} className="flex items-center gap-1">
-              <span className="text-muted-foreground">/</span>
-              <button
-                onClick={() => {
-                  const sub = arr.slice(0, i + 1).join('/');
-                  if (isLocalFolder && activeSyncedFolder) {
-                    fetchLocalFiles(activeSyncedFolder.id, sub);
-                  } else {
-                    // Send the storage-relative path so the API resolver anchors it
-                    // under the drive root (server normalizes a leading drive root).
-                    const fullPath = driveRoot ? `${driveRoot}/${sub}` : sub;
-                    navigateToPath(fullPath);
-                  }
-                }}
-                className={cn(
-                  i === arr.length - 1 ? 'text-foreground' : 'text-muted-foreground hover:text-foreground'
-                )}
-              >
-                {part}
-              </button>
-            </span>
-          ))}
-        </div>
-
-        {/* Toolbar */}
-        <div className="flex items-center justify-between border-b px-4 py-2">
-          <div className="flex items-center gap-2">
-            {/* Hide create/upload buttons for local folders (read-only view) */}
-            {!isLocalFolder && (
-              <>
-                <button
-                  onClick={handleNewFile}
-                  className="flex items-center gap-2 rounded-md bg-primary px-3 py-1.5 text-sm font-medium text-primary-foreground hover:bg-primary/90"
-                >
-                  <FileCode size={14} />
-                  New File
-                </button>
-                <button
-                  onClick={handleNewFolder}
-                  className="flex items-center gap-2 rounded-md border px-3 py-1.5 text-sm hover:bg-muted"
-                >
-                  <FolderPlus size={14} />
-                  New Folder
-                </button>
-                <button
-                  onClick={() => fileInputRef.current?.click()}
-                  className="flex items-center gap-2 rounded-md border px-3 py-1.5 text-sm hover:bg-muted"
-                >
-                  <Upload size={14} />
-                  Upload
-                </button>
-              </>
-            )}
-            <button
-              onClick={handleRefresh}
-              disabled={loading}
-              className={cn(
-                "flex items-center gap-2 rounded-md border px-3 py-1.5 text-sm hover:bg-muted",
-                loading && "opacity-50"
-              )}
-            >
-              <RefreshCw size={14} className={loading ? "animate-spin" : ""} />
-              Refresh
-            </button>
-
-            {/* View toggle */}
-            <div className="flex items-center rounded-md border">
-              <button
-                onClick={() => setViewMode('list')}
-                className={cn(
-                  "flex h-8 w-8 items-center justify-center rounded-l-md",
-                  viewMode === 'list' ? 'bg-muted' : 'hover:bg-muted/50'
-                )}
-              >
-                <List size={14} />
-              </button>
-              <button
-                onClick={() => setViewMode('grid')}
-                className={cn(
-                  "flex h-8 w-8 items-center justify-center rounded-r-md border-l",
-                  viewMode === 'grid' ? 'bg-muted' : 'hover:bg-muted/50'
-                )}
-              >
-                <Grid size={14} />
-              </button>
-            </div>
-            {/* Hidden file input */}
-            <input
-              ref={fileInputRef}
-              type="file"
-              multiple
-              onChange={handleFileInputChange}
-              className="hidden"
-            />
-          </div>
-
-          <div className="flex items-center gap-2">
-            {/* Search */}
-            <div className="relative">
-              <Search size={14} className="absolute left-2.5 top-1/2 -translate-y-1/2 text-muted-foreground" />
-              <input
-                type="text"
-                placeholder="Search files..."
-                value={searchQuery}
-                onChange={(e) => setSearchQuery(e.target.value)}
-                className="h-8 w-48 rounded-md border bg-transparent pl-8 pr-3 text-sm outline-none focus:ring-1 focus:ring-primary"
-              />
-            </div>
-          </div>
-        </div>
-
-        {/* Selection action bar (list view) */}
-        {viewMode === 'list' && !isLocalFolder && selectedFiles.length > 0 && (
-          <div className="flex items-center justify-between border-b bg-muted/30 px-4 py-2 text-sm">
-            <span className="font-medium">
-              {selectedFiles.length} selected
-            </span>
-            <div className="flex items-center gap-2">
-              <button
-                onClick={() => setSelectedFiles([])}
-                className="rounded-md border px-3 py-1.5 text-sm hover:bg-muted"
-              >
-                Clear
-              </button>
-              <button
-                onClick={handleDeleteSelected}
-                className="flex items-center gap-2 rounded-md bg-destructive px-3 py-1.5 text-sm font-medium text-destructive-foreground hover:bg-destructive/90"
-              >
-                <Trash2 size={14} />
-                Delete
-              </button>
-            </div>
-          </div>
-        )}
-
-        {/* Content */}
-        <div 
-          className={cn(
-            "relative flex-1 overflow-auto p-4",
-            isDragging && "bg-primary/5"
-          )}
-          onDragOver={handleDragOver}
-          onDragLeave={handleDragLeave}
-          onDrop={handleDrop}
-        >
-          {/* Drag overlay */}
-          {isDragging && (
-            <div className="absolute inset-0 z-50 flex items-center justify-center bg-background/80 backdrop-blur-sm">
-              <div className="flex flex-col items-center gap-2 rounded-lg border-2 border-dashed border-primary p-8">
-                <Upload size={48} className="text-primary" />
-                <p className="text-lg font-medium">Drop files here to upload</p>
-                <p className="text-sm text-muted-foreground">Files and folders will be uploaded to the current folder</p>
-              </div>
-            </div>
-          )}
-
-          {uploadProgress && (
-            <div className="fixed bottom-4 right-4 z-50 w-80 rounded-lg border bg-background shadow-lg">
-              <div className="flex items-center justify-between border-b px-4 py-2">
-                <div className="flex items-center gap-2">
-                  <Upload size={16} className={uploadProgress.active ? 'animate-pulse text-primary' : 'text-muted-foreground'} />
-                  <p className="text-sm font-medium">
-                    {uploadProgress.active
-                      ? `Uploading ${uploadProgress.completed + 1} of ${uploadProgress.total}`
-                      : uploadProgress.failed > 0
-                        ? `Uploaded ${uploadProgress.completed} of ${uploadProgress.total} (${uploadProgress.failed} failed)`
-                        : `Uploaded ${uploadProgress.completed} file${uploadProgress.completed === 1 ? '' : 's'}`}
-                  </p>
-                </div>
-              </div>
-              <div className="px-4 py-3">
-                <div className="h-1.5 w-full overflow-hidden rounded-full bg-muted">
-                  <div
-                    className="h-full bg-primary transition-all"
-                    style={{
-                      width: `${uploadProgress.total === 0 ? 0 : ((uploadProgress.completed + uploadProgress.failed) / uploadProgress.total) * 100}%`,
-                    }}
-                  />
-                </div>
-                {uploadProgress.currentName && (
-                  <p className="mt-2 truncate text-xs text-muted-foreground" title={uploadProgress.currentName}>
-                    {uploadProgress.currentName}
-                  </p>
-                )}
-                {uploadProgress.active && (
-                  <p className="mt-2 text-xs text-muted-foreground">
-                    Please keep this page open until the upload finishes.
-                  </p>
-                )}
-              </div>
-            </div>
-          )}
-
-          {error && (
-            <div className="mb-4 rounded-md bg-destructive/10 p-3 text-sm text-destructive">
-              {error}
-            </div>
-          )}
-
-          {filteredFiles.length === 0 ? (
-            <div className="flex flex-col items-center justify-center py-16 text-center">
-              <div className="mb-4 flex h-16 w-16 items-center justify-center rounded-2xl bg-muted">
-                <Folder size={32} className="text-muted-foreground" />
-              </div>
-              <h3 className="mb-2 text-lg font-medium">
-                {searchQuery ? 'No files found' : isLocalFolder ? 'Folder is empty' : 'No files yet'}
-              </h3>
-              <p className="mb-4 text-muted-foreground">
-                {searchQuery 
-                  ? 'Try a different search term' 
-                  : isLocalFolder 
-                    ? 'This folder has no files or subfolders'
-                    : 'Create a file or folder to get started'}
-              </p>
-              {!searchQuery && !isLocalFolder && (
-                <div className="flex flex-col items-center gap-4">
-                  <div className="flex gap-2">
-                    <button
-                      onClick={handleNewFile}
-                      className="flex items-center gap-2 rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground"
-                    >
-                      <FileCode size={14} />
-                      New File
-                    </button>
-                    <button
-                      onClick={handleNewFolder}
-                      className="flex items-center gap-2 rounded-md border px-4 py-2 text-sm"
-                    >
-                      <FolderPlus size={14} />
-                      New Folder
-                    </button>
-                    <button
-                      onClick={() => fileInputRef.current?.click()}
-                      className="flex items-center gap-2 rounded-md border px-4 py-2 text-sm"
-                    >
-                      <Upload size={14} />
-                      Upload Files
-                    </button>
-                  </div>
-                  <p className="text-xs text-muted-foreground">
-                    Or drag and drop files anywhere on this page
-                  </p>
-                </div>
-              )}
-            </div>
-          ) : viewMode === 'list' ? (
-            /* List view */
-            <table className="w-full">
-              <thead>
-                <tr className="border-b text-left text-xs text-muted-foreground">
-                  {!isLocalFolder && (
-                    <th className="w-8 pb-2 pr-2">
-                      <input
-                        type="checkbox"
-                        aria-label="Select all"
-                        checked={allSelected}
-                        ref={(el) => {
-                          if (el) el.indeterminate = someSelected;
-                        }}
-                        onChange={toggleSelectAll}
-                        className="h-4 w-4 cursor-pointer rounded border-border accent-primary"
-                      />
-                    </th>
-                  )}
-                  <th className="pb-2 font-medium">
-                    <SortHeader column="name" label="Name" />
-                  </th>
-                  <th className="pb-2 font-medium">
-                    <SortHeader column="size" label="Size" />
-                  </th>
-                  <th className="pb-2 font-medium">
-                    <SortHeader column="modified" label="Modified" />
-                  </th>
-                  <th className="pb-2 font-medium w-16"></th>
-                </tr>
-              </thead>
-              <tbody>
-                {pagedFiles.map((file) => (
-                  <tr
-                    key={file.path}
-                    draggable={!isLocalFolder}
-                    onDragStart={(e) => handleRowDragStart(e, file)}
-                    onDragOver={file.type === 'folder' && !isLocalFolder ? (e) => handleFolderDragOver(e, file) : undefined}
-                    onDragLeave={file.type === 'folder' && !isLocalFolder ? (e) => handleFolderDragLeave(e, file) : undefined}
-                    onDrop={file.type === 'folder' && !isLocalFolder ? (e) => handleFolderDrop(e, file) : undefined}
-                    className={cn(
-                      "group border-b border-border/50 hover:bg-muted/50",
-                      selectedFiles.includes(file.path) && "bg-primary/5",
-                      dropTargetPath === file.path && "bg-primary/10 ring-1 ring-primary"
-                    )}
-                  >
-                    {!isLocalFolder && (
-                      <td className="w-8 py-2 pr-2">
-                        <input
-                          type="checkbox"
-                          aria-label={`Select ${file.name}`}
-                          checked={selectedFiles.includes(file.path)}
-                          onChange={() => toggleSelectFile(file.path)}
-                          className="h-4 w-4 cursor-pointer rounded border-border accent-primary"
-                        />
-                      </td>
-                    )}
-                    <td className="py-2">
-                      <button
-                        onClick={() => {
-                          if (file.type === 'folder') {
-                            if (isLocalFolder && activeSyncedFolder) {
-                              fetchLocalFiles(activeSyncedFolder.id, file.path);
-                            } else {
-                              navigateToPath(file.path);
-                            }
-                          } else if (isOfficeFile(file)) {
-                            openOfficePreview(file);
-                          } else if (isImageFile(file)) {
-                            openImageViewer(file);
-                          } else if (isPdfFile(file)) {
-                            openPdfViewer(file);
-                          } else if (isTextFile(file)) {
-                            openTextViewer(file);
-                          } else {
-                            openFile(file.path);
-                          }
-                        }}
-                        className="flex items-center gap-2 text-sm hover:text-primary"
-                      >
-                        {getFileIcon(file, 16)}
-                        {file.name}
-                      </button>
-                    </td>
-                    <td className="py-2 text-sm text-muted-foreground">
-                      {file.type === 'folder' ? '—' : formatSize(file.size)}
-                    </td>
-                    <td className="py-2 text-sm text-muted-foreground">
-                      {formatDate(file.modified)}
-                    </td>
-                    <td className="py-2">
-                      <div className="relative flex items-center justify-end gap-0.5">
-                        {(() => {
-                          const starred = starredItems.some(
-                            (i) => i.path === file.path && i.workspaceId === workspaceId
-                          );
-                          return (
-                            <button
-                              title={starred ? 'Remove from starred' : 'Add to starred'}
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                if (starred) {
-                                  unstarItem(file.path, workspaceId);
-                                } else {
-                                  starItem({
-                                    path: file.path,
-                                    name: file.name,
-                                    type: file.type,
-                                    source: activeSource,
-                                    workspaceId,
-                                  });
-                                }
-                              }}
-                              className={cn(
-                                'h-6 w-6 items-center justify-center rounded hover:bg-muted',
-                                starred
-                                  ? 'flex text-amber-400 hover:text-amber-500'
-                                  : 'hidden text-muted-foreground/50 hover:text-foreground group-hover:flex'
-                              )}
-                            >
-                              <Star size={14} className={starred ? 'fill-current' : ''} />
-                            </button>
-                          );
-                        })()}
-                        <button
-                          title={file.type === 'folder' ? 'Download as ZIP' : 'Download'}
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            if (file.type === 'folder') {
-                              downloadFolderToDesktop(file);
-                            } else {
-                              downloadFileToDesktop(file);
-                            }
-                          }}
-                          className="hidden h-6 w-6 items-center justify-center rounded text-muted-foreground/50 hover:bg-muted hover:text-foreground group-hover:flex"
-                        >
-                          <Download size={14} />
-                        </button>
-                        <button
-                          onClick={(e) => {
-                            e.stopPropagation();
-                            setActiveContextMenu(activeContextMenu === file.path ? null : file.path);
-                          }}
-                          className="flex h-6 w-6 items-center justify-center rounded text-muted-foreground/50 hover:bg-muted hover:text-foreground group-hover:text-muted-foreground"
-                        >
-                          <MoreVertical size={14} />
-                        </button>
-                        {activeContextMenu === file.path && (
-                          <div className="absolute right-0 top-full z-50 min-w-[140px] rounded-md border bg-popover py-1 shadow-lg">
-                            {file.type === 'file' && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  handleOpenInLab(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <FlaskConical size={12} />
-                                Open in Lab
-                              </button>
-                            )}
-                            {isOfficeFile(file) && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  openOfficePreview(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <Eye size={12} />
-                                Preview {officeKindLabel(file)}
-                              </button>
-                            )}
-                            {isPdfFile(file) && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  openPdfViewer(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <Eye size={12} />
-                                View PDF
-                              </button>
-                            )}
-                            {isImageFile(file) && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  openImageViewer(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <Eye size={12} />
-                                View Image
-                              </button>
-                            )}
-                            {isTextFile(file) && !isPdfFile(file) && !isImageFile(file) && !isOfficeFile(file) && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  openTextViewer(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <Eye size={12} />
-                                {isMarkdownFile(file) ? 'View Markdown' : 'Preview'}
-                              </button>
-                            )}
-                            {file.type === 'file' && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  downloadFileToDesktop(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <Download size={12} />
-                                Download
-                              </button>
-                            )}
-                            {file.type === 'folder' && (
-                              <button
-                                onClick={(e) => {
-                                  e.stopPropagation();
-                                  setActiveContextMenu(null);
-                                  downloadFolderToDesktop(file);
-                                }}
-                                className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                              >
-                                <Download size={12} />
-                                Download as ZIP
-                              </button>
-                            )}
-                            <button
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                setActiveContextMenu(null);
-                                handleRename(file);
-                              }}
-                              className="flex w-full items-center gap-2 px-3 py-1.5 text-xs hover:bg-muted"
-                            >
-                              Rename
-                            </button>
-                            <button
-                              onClick={(e) => {
-                                e.stopPropagation();
-                                setActiveContextMenu(null);
-                                handleDelete(file);
-                              }}
-                              className="flex w-full items-center gap-2 px-3 py-1.5 text-xs text-destructive hover:bg-destructive/10"
-                            >
-                              Delete
-                            </button>
-                          </div>
-                        )}
-                      </div>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          ) : (
-            /* Grid view */
-            <div className="grid grid-cols-2 gap-4 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-6">
-              {pagedFiles.map((file) => (
-                <div
-                  key={file.path}
-                  draggable={!isLocalFolder}
-                  onDragStart={(e) => handleRowDragStart(e, file)}
-                  onDragOver={file.type === 'folder' && !isLocalFolder ? (e) => handleFolderDragOver(e, file) : undefined}
-                  onDragLeave={file.type === 'folder' && !isLocalFolder ? (e) => handleFolderDragLeave(e, file) : undefined}
-                  onDrop={file.type === 'folder' && !isLocalFolder ? (e) => handleFolderDrop(e, file) : undefined}
-                  className={cn(
-                    "group relative rounded-lg",
-                    dropTargetPath === file.path && "ring-2 ring-primary"
-                  )}
-                >
-                  <button
-                    onClick={() => {
-                      if (file.type === 'folder') {
-                        if (isLocalFolder && activeSyncedFolder) {
-                          fetchLocalFiles(activeSyncedFolder.id, file.path);
-                        } else {
-                          navigateToPath(file.path);
-                        }
-                      } else if (isOfficeFile(file)) {
-                        openOfficePreview(file);
-                      } else if (isImageFile(file)) {
-                        openImageViewer(file);
-                      } else if (isPdfFile(file)) {
-                        openPdfViewer(file);
-                      } else if (isTextFile(file)) {
-                        openTextViewer(file);
-                      } else {
-                        openFile(file.path);
-                      }
-                    }}
-                    className="flex w-full flex-col items-center gap-2 rounded-lg border p-4 hover:bg-muted/50"
-                  >
-                    {getFileIcon(file, 40)}
-                    <span className="w-full truncate text-center text-sm">{file.name}</span>
-                  </button>
-                  {(() => {
-                    const starred = starredItems.some(
-                      (i) => i.path === file.path && i.workspaceId === workspaceId
-                    );
-                    return (
-                      <button
-                        title={starred ? 'Remove from starred' : 'Add to starred'}
-                        onClick={(e) => {
-                          e.stopPropagation();
-                          if (starred) {
-                            unstarItem(file.path, workspaceId);
-                          } else {
-                            starItem({
-                              path: file.path,
-                              name: file.name,
-                              type: file.type,
-                              source: activeSource,
-                              workspaceId,
-                            });
-                          }
-                        }}
-                        className={cn(
-                          'absolute left-1.5 top-1.5 h-6 w-6 items-center justify-center rounded bg-background/80 shadow-sm',
-                          starred
-                            ? 'flex text-amber-400 hover:text-amber-500'
-                            : 'hidden text-muted-foreground hover:bg-muted hover:text-foreground group-hover:flex'
-                        )}
-                      >
-                        <Star size={13} className={starred ? 'fill-current' : ''} />
-                      </button>
-                    );
-                  })()}
-                  <button
-                    title={file.type === 'folder' ? 'Download as ZIP' : 'Download'}
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      if (file.type === 'folder') {
-                        downloadFolderToDesktop(file);
-                      } else {
-                        downloadFileToDesktop(file);
-                      }
-                    }}
-                    className="absolute right-1.5 top-1.5 hidden h-6 w-6 items-center justify-center rounded bg-background/80 text-muted-foreground shadow-sm hover:bg-muted hover:text-foreground group-hover:flex"
-                  >
-                    <Download size={13} />
-                  </button>
-                </div>
-              ))}
-            </div>
-          )}
-        </div>
-
-        {/* Pagination bar */}
-        {pageCount > 0 && (
-          <div className="flex flex-wrap items-center justify-between gap-2 border-t px-4 py-2 text-sm">
-            <div className="flex items-center gap-2 text-muted-foreground">
-              <span>Rows per page</span>
-              <select
-                value={pageSize}
-                onChange={(e) => changePageSize(Number(e.target.value))}
-                className="h-8 rounded-md border bg-transparent px-2 outline-none focus:ring-1 focus:ring-primary"
-              >
-                {PAGE_SIZE_OPTIONS.map((size) => (
-                  <option key={size} value={size}>
-                    {size}
-                  </option>
-                ))}
-              </select>
-            </div>
-            <div className="flex items-center gap-4">
-              <span className="text-muted-foreground">
-                {`${rangeStart}–${rangeEnd} of ${pageCount}`}
-              </span>
-              <div className="flex items-center gap-1">
-                <button
-                  onClick={() => goToPage(safePage - 1)}
-                  disabled={loading || safePage <= 1}
-                  className="rounded-md border px-3 py-1.5 hover:bg-muted disabled:opacity-50 disabled:hover:bg-transparent"
-                >
-                  Previous
-                </button>
-                <span className="px-2 text-muted-foreground">
-                  {`Page ${safePage} of ${totalPages}`}
-                </span>
-                <button
-                  onClick={() => goToPage(safePage + 1)}
-                  disabled={loading || safePage >= totalPages}
-                  className="rounded-md border px-3 py-1.5 hover:bg-muted disabled:opacity-50 disabled:hover:bg-transparent"
-                >
-                  Next
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-
-    </div>
-
-      {/* PDF Viewer modal — rendered outside the page container to avoid stacking-context traps */}
-      {pdfViewerFileName && (
-        <div className="fixed inset-0 z-[200] bg-black/60 p-4 backdrop-blur-sm">
-          <div className="flex h-full flex-col overflow-hidden rounded-xl border bg-background shadow-2xl">
-            <div className="flex items-center justify-between border-b px-4 py-2">
-              <div className="truncate text-sm font-medium">{pdfViewerFileName}</div>
-              <div className="flex items-center gap-2">
-                {pdfViewerDownloadFile && (
-                  <button
-                    onClick={() => downloadFileToDesktop(pdfViewerDownloadFile)}
-                    className="flex items-center gap-2 rounded-md border px-3 py-1.5 text-xs hover:bg-muted"
-                  >
-                    <Download size={12} />
-                    Download original
-                  </button>
-                )}
-                <button
-                  onClick={closePdfViewer}
-                  className="flex h-8 w-8 items-center justify-center rounded-md hover:bg-muted"
-                  aria-label="Close PDF viewer"
-                >
-                  <X size={16} />
-                </button>
-              </div>
-            </div>
-            <div className="relative flex-1 bg-muted/20">
-              {pdfViewerLoading && (
-                <div className="absolute inset-0 flex items-center justify-center text-sm text-muted-foreground">
-                  Loading PDF preview...
-                </div>
-              )}
-              {pdfViewerError && !pdfViewerLoading && (
-                <div className="absolute inset-0 flex items-center justify-center p-6 text-sm text-destructive">
-                  {pdfViewerError}
-                </div>
-              )}
-              {pdfViewerUrl && !pdfViewerLoading && !pdfViewerError && (
-                <PdfViewer src={pdfViewerUrl} className="h-full w-full" />
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-      {imageViewerFileName && (
-        <div className="fixed inset-0 z-[200] bg-black/60 p-4 backdrop-blur-sm">
-          <div className="flex h-full flex-col overflow-hidden rounded-xl border bg-background shadow-2xl">
-            <div className="flex items-center justify-between border-b px-4 py-2">
-              <div className="truncate text-sm font-medium">{imageViewerFileName}</div>
-              <button
-                onClick={closeImageViewer}
-                className="flex h-8 w-8 items-center justify-center rounded-md hover:bg-muted"
-                aria-label="Close image viewer"
-              >
-                <X size={16} />
-              </button>
-            </div>
-            <div className="relative flex-1 bg-black/40">
-              {imageViewerLoading && (
-                <div className="absolute inset-0 flex items-center justify-center text-sm text-white/80">
-                  Loading image preview...
-                </div>
-              )}
-              {imageViewerError && !imageViewerLoading && (
-                <div className="absolute inset-0 flex items-center justify-center p-6 text-sm text-destructive">
-                  {imageViewerError}
-                </div>
-              )}
-              {imageViewerUrl && !imageViewerLoading && !imageViewerError && (
-                <div className="flex h-full w-full items-center justify-center p-4">
-                  {/* eslint-disable-next-line @next/next/no-img-element */}
-                  <img
-                    src={imageViewerUrl}
-                    alt={imageViewerFileName}
-                    className="max-h-full max-w-full object-contain"
-                  />
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-      {textViewerFileName && (
-        <div className="fixed inset-0 z-[200] bg-black/60 p-4 backdrop-blur-sm">
-          <div className="flex h-full flex-col overflow-hidden rounded-xl border bg-background shadow-2xl">
-            <div className="flex items-center justify-between border-b px-4 py-2">
-              <div className="flex min-w-0 items-center gap-2">
-                {textViewerIsHtml && (
-                  <div className="flex items-center rounded-md border">
-                    <button
-                      onClick={() => setHtmlPreview(true)}
-                      title="Preview"
-                      aria-label="Preview rendered HTML"
-                      className={cn(
-                        'flex h-7 w-7 items-center justify-center rounded-l-md',
-                        htmlPreview ? 'bg-muted' : 'hover:bg-muted/50'
-                      )}
-                    >
-                      <Eye size={14} />
-                    </button>
-                    <button
-                      onClick={() => setHtmlPreview(false)}
-                      title="View source"
-                      aria-label="View HTML source"
-                      className={cn(
-                        'flex h-7 w-7 items-center justify-center rounded-r-md border-l',
-                        !htmlPreview ? 'bg-muted' : 'hover:bg-muted/50'
-                      )}
-                    >
-                      <Code size={14} />
-                    </button>
-                  </div>
-                )}
-                <div className="truncate text-sm font-medium">{textViewerFileName}</div>
-              </div>
-              <button
-                onClick={closeTextViewer}
-                className="flex h-8 w-8 items-center justify-center rounded-md hover:bg-muted"
-                aria-label="Close text viewer"
-              >
-                <X size={16} />
-              </button>
-            </div>
-            <div className="relative flex-1 overflow-auto bg-background">
-              {textViewerLoading && (
-                <div className="absolute inset-0 flex items-center justify-center text-sm text-muted-foreground">
-                  Loading preview...
-                </div>
-              )}
-              {textViewerError && !textViewerLoading && (
-                <div className="absolute inset-0 flex items-center justify-center p-6 text-sm text-destructive">
-                  {textViewerError}
-                </div>
-              )}
-              {!textViewerLoading && !textViewerError && textViewerIsHtml && htmlPreview && (
-                <iframe
-                  title={textViewerFileName ?? 'HTML preview'}
-                  srcDoc={textViewerContent}
-                  sandbox="allow-scripts allow-popups allow-forms allow-modals"
-                  className="h-full w-full border-0 bg-white"
-                />
-              )}
-              {!textViewerLoading && !textViewerError && textViewerIsHtml && !htmlPreview && (
-                <pre className="m-0 h-full overflow-auto whitespace-pre p-4 font-mono text-xs leading-relaxed text-foreground">
-                  {textViewerContent}
-                </pre>
-              )}
-              {!textViewerLoading && !textViewerError && !textViewerIsHtml && textViewerMode === 'markdown' && (
-                <div className="prose prose-sm max-w-none p-6 dark:prose-invert">
-                  <ReactMarkdown remarkPlugins={[remarkGfm]}>
-                    {textViewerContent}
-                  </ReactMarkdown>
-                </div>
-              )}
-              {!textViewerLoading && !textViewerError && !textViewerIsHtml && textViewerMode === 'code' && (
-                <pre className="m-0 h-full overflow-auto whitespace-pre p-4 font-mono text-xs leading-relaxed text-foreground">
-                  {textViewerContent}
-                </pre>
-              )}
-            </div>
-          </div>
-        </div>
-      )}
-    </>
-  );
+import FilesBrowse from './browse/browse';
+import { useIsMobile } from '@/hooks/use-is-mobile';
+
+/**
+ * Desktop: /files is the file browser.
+ * Mobile: /files is the drive list (workspace-layout renders FilesSection);
+ * do not mount the browser here or its fetch effects fight the shell.
+ */
+export default function FilesIndexPage() {
+  const isMobile = useIsMobile();
+  if (isMobile) return null;
+  return <FilesBrowse />;
 }
